requirements.txt
# Python 3.7.6
Django
djangorestframework
djangorestframework-gis
psycopg2
Pillow
drf-yasg
django-jazzmin